Extract Loader type into separate source file
This change moves the Loader type from sat.go into loader.go so as to
make it a more distinct unit that can be tested directly.
